% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_model.R
\name{boot_model}
\alias{boot_model}
\title{boot_model}
\usage{
boot_model(data, outcome, boot_reps, model, ...)
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats
}
