% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_enet.R
\name{model_enet}
\alias{model_enet}
\title{model_enet}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted (usually coming from nested bootstrap data)}

\item{outcome}{the outcome as a string (i.e. "y")}

\item{type}{model type, either "linear" or "logistic"}
}
\description{
Function to model elastic net selection process on a given dataframe
}
