% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lasso.R
\name{model_lasso}
\alias{model_lasso}
\title{model_lasso}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted (usually coming from nested bootstrap data)}

\item{outcome}{the outcome as a string (i.e. "y")}
}
\description{
Function to model lasso selection process on a given dataframe
}
