% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mbic.R
\name{model_mbic}
\alias{model_mbic}
\title{model_mbic}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted (usually coming from nested bootstrap data)}

\item{outcome}{the outcome as a string (i.e. "y")}
}
\description{
Function to model mbic selection process on a given dataframe
}
