% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mcp.R
\name{model_mcp}
\alias{model_mcp}
\title{model_mcp}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted (usually coming from nested bootstrap data)}

\item{outcome}{the outcome as a string (i.e. "y")}
}
\description{
Function to model mcp selection process on a given dataframe
}
