% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilise.R
\name{model_selector}
\alias{model_selector}
\title{stabilise}
\usage{
model_selector(selected_model)
}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted}

\item{outcome}{the outcome as a string (i.e. "y")}

\item{boot_reps}{the number of bootstrap samples}

\item{permutations}{the number of times to be permuted per repeat}

\item{perm_boot_reps}{the number of times to repeat each set of permutations}
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats
}
