% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_stab.R
\name{perm_stab}
\alias{perm_stab}
\title{model_selector}
\usage{
perm_stab(data, outcome, boot_reps, permutations, perm_boot_reps, model_name)
}
\description{
Main function to call both permutation and bootstrapping functions; to be looped over multiple models selected by the user.
}
