% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_selector.R
\name{rep_selector_boot}
\alias{rep_selector_boot}
\title{rep_selector_boot}
\usage{
rep_selector_boot(data, boot_reps)
}
\arguments{
\item{data}{the dataset to analyse.}

\item{boot_reps}{the number of bootstrap samples}
}
\description{
wrapper function to determine the number of bootstrap repeats
}
