% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_selector.R
\name{rep_selector_perm}
\alias{rep_selector_perm}
\title{rep_selector_boot}
\usage{
rep_selector_perm(data, permutations)
}
\arguments{
\item{data}{the dataset to analyse.}

\item{permutations}{the number of times to be permuted per repeat}
}
\description{
wrapper function to determine the number of permutations
}
\keyword{internal}
