% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_bias_inner.R
\name{selection_bias_inner}
\alias{selection_bias_inner}
\title{selection_bias_inner}
\arguments{
\item{nrows}{The number of rows to simulate.}

\item{ncols}{The number of columns to simulate.}

\item{p_thresh}{The p-value threshold to use in univariate pre-filtration.}
}
\value{
A list including a dataframe of results, a dataframe of the median number of variables selected and a plot illustrating false positive selection.
}
\description{
An function to illustrate the risk of selection bias in
conventional modelling approaches by simulating a dataset with no information and
conducting conventional modelling with prefiltration.
}
