% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(nrows, ncols, n_true = 0, amplitude = 0)
}
\arguments{
\item{nrows}{The number of rows to simulate.}

\item{ncols}{The number of columns to simulate.}

\item{n_true}{The number of variables truly associated with the outcome.}

\item{amplitude}{The strength of association between true variables and the outcome.}
}
\value{
A simulated dataset
}
\description{
Simulate a dataset. This can optionally include variables with a given associated with the outcome.
}
