% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_re.R
\name{simulate_data_re}
\alias{simulate_data_re}
\title{simulate_data_re}
\usage{
simulate_data_re(sd_level_1 = 2, sd_level_2 = 2)
}
\arguments{
\item{sd_level_1}{Standard deviation of level 1 variables}

\item{sd_level_2}{Standard deviation of level 2 variables}
}
\value{
A simulated dataset
}
\description{
Simulate a 500x500 dataset with random effects. This can optionally include variables with a given associated with the outcome.
}
