% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_glmer_re.R
\name{simulate_data_re_glmer}
\alias{simulate_data_re_glmer}
\alias{simulate_glmer_re_data}
\title{simulate_data_re_glmer}
\usage{
simulate_glmer_re_data(
  n_subjects = 100,
  obs_per_subject = 10,
  n_signal = 2,
  n_noise = 3,
  beta0 = -1,
  beta_signal = NULL,
  sigma_u = 1
)
}
\arguments{
\item{n_subjects}{The number of individual subjects, e.g. participations}

\item{obs_per_subject}{The number of observations per subject}

\item{n_signal}{The  number of causal predictors}

\item{n_noise}{The number of junk predictors}

\item{beta0}{Intercept}

\item{beta_signal}{signal size for causal parameters}

\item{sigma_u}{standard deviation for random intercepts}
}
\value{
A simulated dataset with a clustered outcome suitable for random effects analysis with a binary outcome
}
\description{
Simulate a dataset where some variables are associated with the outcome and some are unk
}
