% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_selection_bias.R
\name{simulate_selection_bias}
\alias{simulate_selection_bias}
\title{simulate_selection_bias}
\arguments{
\item{nrows}{A vector of the number of rows to simulate (i.e., c(100, 200)).}

\item{ncols}{A vector of the number of columns to simulate (i.e., c(100, 200)).}

\item{p_thresh}{A vector of the p-value threshold to use in univariate pre-filtration (i.e., c(0.1, 0.2)).}
}
\value{
A list including a dataframe of results, a dataframe of the median number of variables selected and a plot illustrating false positive selection.
}
\description{
An function to illustrate the risk of selection bias in
conventional modelling approaches by simulating a dataset with no information and
conducting conventional modelling with prefiltration.
}
