% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_plot.R
\name{stab_boot_plot}
\alias{stab_boot_plot}
\title{stab_plot}
\usage{
stab_boot_plot(model)
}
\arguments{
\item{stabiliser_outcome}{outcome from stabiliser() function}
}
\description{
Plot from stability object
}
