% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_plot.R
\name{stab_plot}
\alias{stab_plot}
\title{stab_plot}
\arguments{
\item{stabiliser_outcome}{Outcome from stabilise() or triangulate() function.}
}
\value{
A ggplot object.
}
\description{
Plot from stability object
}
