% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilise.R
\name{stabilise}
\alias{stabilise}
\title{stabilise}
\usage{
stabilise(data, outcome, boot_reps, permutations, perm_boot_reps, model)
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats
}
