% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilise_re.R
\name{stabilise_re}
\alias{stabilise_re}
\title{stabilise_re}
\arguments{
\item{data}{A dataframe containing an outcome variable to be permuted.}

\item{outcome}{The outcome as a string (i.e. "y").}

\item{level_2_id}{The variable name determining level 2 status as a string (i.e., "level_2_column_name").}

\item{n_top_filter}{The number of variables to filter for final model (Default = 50).}

\item{boot_reps}{The number of bootstrap samples. Default is "auto" which selects number based on dataframe size.}

\item{permutations}{The number of times to be permuted per repeat. Default is "auto" which selects number based on dataframe size.}

\item{perm_boot_reps}{The number of times to repeat each set of permutations. Default is 5.}

\item{normalise}{Normalise numeric variables (TRUE/FALSE)}

\item{dummy}{Create dummy variables for factors/characters (TRUE/FALSE)}

\item{impute}{Impute missing data (TRUE/FALSE)}
}
\value{
A list containing a table of variable stabilities and a numeric permutation threshold.
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats using clustered data.
}
