% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilise_re_glmer.R
\name{stabilise_re_glmer}
\alias{stabilise_re_glmer}
\title{stabilise_re_glmer}
\arguments{
\item{data}{A dataframe containing an outcome variable to be permuted.}

\item{outcome}{The outcome as a string (i.e. "y").}

\item{intercept_level_ids}{A vector names defining which variables are random effect, i.e., c("level_2_column_name", "level_3_column_name").}

\item{n_top_filter}{The number of variables to filter for final model (Default = 50).}

\item{boot_reps}{The number of bootstrap samples. Default is "auto" which selects number based on dataframe size. For glmer models, these are subsamples of the dataset, set to 80\%.}

\item{permutations}{The number of times to be permuted per repeat. Default is "auto" which selects number based on dataframe size.}

\item{perm_boot_reps}{The number of times to repeat each set of permutations. Default is 20.}

\item{normalise}{Normalise numeric variables (TRUE/FALSE)}

\item{dummy}{Create dummy variables for factors/characters (TRUE/FALSE)}

\item{impute}{Impute missing data (TRUE/FALSE)}

\item{base_id}{level of the random effect to bootstrap by, e.g individual. This is likely the lower level of random effect specified}

\item{parallel}{TRUE or FALSE, whether to set up parallel processing}

\item{num_cores}{Number of cores to use if parallel processing required}
}
\value{
A list containing a table of variable stabilities and a numeric permutation threshold.
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats using clustered data.
}
