% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabiliser_example.R
\docType{data}
\name{stabiliser_example}
\alias{stabiliser_example}
\title{stabiliser_example}
\format{
A data frame with 50 rows and 100 variables.
The stabiliser_example dataset is a simulated example with the following properties:
1 simulated outcome variable: \code{y}
4 variables simulated to be associated with \code{y}: \code{causal1}, \code{causal2}...
95 variables simulated to have no association with \code{y}: \code{junk1}, \code{junk2}...
}
\usage{
stabiliser_example
}
\description{
A simulated dataset
}
\keyword{datasets}
