% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{stabiliser_prep}
\alias{stabiliser_prep}
\alias{prep_data}
\title{stabiliser_prep}
\usage{
prep_data(outcome, data, normalise, dummy, impute)
}
\arguments{
\item{normalise}{Normalise numeric variables (TRUE/FALSE)}

\item{dummy}{Create dummy variables for factors/characters (TRUE/FALSE)}

\item{impute}{Impute missing data (TRUE/FALSE)}
}
\description{
Prepares dataset using recipes framework
}
\keyword{internal}
