% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_model.R
\name{boot_model}
\alias{boot_model}
\title{boot_model}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted}

\item{outcome}{the outcome as a string (i.e. "y")}

\item{boot_reps}{the number of bootstrap samples}

\item{model}{the model to be used (i.e. model_mbic)}
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats
}
