% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Permute}
\usage{
permute(data, outcome, permutations, perm_boot_reps, model)
}
\arguments{
\item{data}{a dataframe containing an outcome variable to be permuted}

\item{outcome}{the outcome to be permuted as a string (i.e. "y")}

\item{permutations}{the number of times to be permuted per repeat}

\item{perm_boot_reps}{the number of times to repeat each set of permutations}

\item{model}{the model to be used (i.e. model_mbic)}
}
\description{
Calculates permutation threshold for null model, where a specified model is run over multiple bootstrap resamples of multiple permuted version of the dataset.
}
