% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_re.R
\name{simulate_data_re}
\alias{simulate_data_re}
\alias{simulate_data}
\title{simulate_data_re}
\usage{
simulate_data(
  nrows,
  ncols,
  n_true = 2,
  amplitude = 2,
  n_level_2 = 2,
  sd_level_1 = 0.5,
  sd_level_2 = 0
)
}
\arguments{
\item{nrows}{The number of rows to simulate.}

\item{ncols}{The number of columns to simulate.}

\item{n_true}{The number of variables truly associated with the outcome.}

\item{amplitude}{The strength of association between true variables and the outcome.}
}
\value{
A simulated dataset
}
\description{
Simulate a dataset with random effects. This can optionally include variables with a given associated with the outcome.
}
