% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_re.R
\name{simulate_data_re}
\alias{simulate_data_re}
\title{simulate_data_re}
\arguments{
\item{sd_level_1}{Standard deviation of level 1 variables}

\item{sd_level_2}{Standard deviation of level 2 variables}
}
\value{
A simulated dataset with a clustered outcome sutable for random effects analysis
}
\description{
Simulate a 500x500 dataset with 8 true fixed effects, 492 junk variables and a clustered outcome suitable for a 2 level random effects analysis. The strength of association between true variables and the outcome is governed by the error added at level 1 (defined by parameter sd_level_1) and level 2 (sd_level_2).
}
