% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_plot.R
\name{stab_plot}
\alias{stab_plot}
\title{stab_plot}
\usage{
stab_plot(stabiliser_object)
}
\arguments{
\item{stabiliser_outcome}{outcome from stabiliser() function}
}
\description{
Plot from stability object
}
