% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilise.R
\name{stabilise}
\alias{stabilise}
\title{stabilise}
\arguments{
\item{data}{A dataframe containing an outcome variable to be permuted.}

\item{outcome}{The outcome as a string (i.e. "y").}

\item{boot_reps}{The number of bootstrap samples. Default is "auto" which selects number based on dataframe size.}

\item{permutations}{The number of times to be permuted per repeat. Default is "auto" which selects number based on dataframe size.}

\item{perm_boot_reps}{The number of times to repeat each set of permutations. Default is 5.}

\item{models}{The models to select for stabilising. Default is elastic net (models = c("enet")), other available models include "lasso", "mbic", "mcp".}

\item{type}{The type of model, either "linear" or "logistic"}

\item{quantile}{The quantile of null stabilities to use as a threshold.}

\item{normalise}{Normalise numeric variables (TRUE/FALSE)}

\item{dummy}{Create dummy variables for factors/characters (TRUE/FALSE)}

\item{impute}{Impute missing data (TRUE/FALSE)}
}
\value{
A list for each model selected. Each list contains a dataframe of variable stabilities, a numeric permutation threshold, and a dataframe of coefficients for both bootstrap and permutation.
}
\description{
Function to calculate stability of variables' association with an outcome for a given model over a number of bootstrap repeats
}
