% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabiliser.R
\docType{package}
\name{stabiliser}
\alias{stabiliser}
\alias{stabiliser-package}
\title{stabiliser}
\description{
This package uses bootstrap resampling and an objective selection stability threshold to provide a robust method of selecting variables truly associated with an outcome.
}
\author{
Robert Hyde \email{robert.hyde1@nottingham.ac.uk}

Martin Green

Eliana Lima
}
\keyword{internal}
