% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{stabiliser_prep}
\alias{stabiliser_prep}
\title{stabiliser_prep}
\arguments{
\item{normalise}{Normalise numeric variables (TRUE/FALSE)}

\item{dummy}{Create dummy variables for factors/characters (TRUE/FALSE)}

\item{impute}{Impute missing data (TRUE/FALSE)}
}
\description{
Prepares dataset using recipes framework
}
\keyword{internal}
