% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{triangulate}
\arguments{
\item{object}{an object generated through the stability() function}
}
\description{
Triangulate multiple models using a stability object
}
