% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{triangulate}
\arguments{
\item{object}{An object generated through the stabilise() function.}
}
\value{
A combined list of model results including a dataframe of stability results for variables, and a numeric permutation threshold.
}
\description{
Triangulate multiple models using a stability object
}
